pyproject.toml
[tool.poetry]
name = "qr-generator"
version = "0.1.0"
description = ""
authors = ["kmjm"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "^2.148.0"
aws-cdk-aws-lambda-python-alpha = "^2.148.0a0"
constructs = "^10.3.0"
requests = "^2.32.3"
numpy = "^2.0.0"
boto3 = "^1.34.140"


[tool.poetry.group.test.dependencies]
ruff = "^0.5.1"
[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    "layer"
]


[tool.poetry.group.lambdaDependencies.dependencies]
qrcode = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
